# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Token Swap Workbench API
  version: 0.1.0
  description: HTTP API exposed by the Go workbench in front of the Rust chain service.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /v1/swaps:
    post:
      summary: Submit a token swap
      operationId: submitSwap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwapRequest"
            example:
              account_id: 2
              in_token: NEX
              out_token: ETH
              amount_in: 10
      responses:
        "200":
          description: Swap accepted by the upstream service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
              example:
                status: submitted
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Upstream chain service error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    StatusResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: ok
...
    SwapRequest:
      type: object
      required:
        - account_id
        - in_token
        - out_token
        - amount_in
      properties:
        account_id:
          type: integer
          minimum: 1
          example: 2
        in_token:
          type: string
          example: NEX
        out_token:
          type: string
          example: ETH
        amount_in:
          type: number
          format: double
          exclusiveMinimum: 0
          example: 10
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: BAD_REQUEST
        message:
          type: string
          example: in is required